+++ appregistry/integration/app_test.go
@@ -15,7 +15,7 @@ import (
 	envtest "github.com/ignite/cli/v28/integration"
 )
 
-func TestMarketplace(t *testing.T) {
+func TestAppRegistry(t *testing.T) {
 	var (
 		require = require.New(t)
 		env     = envtest.New(t)
@@ -23,9 +23,9 @@ func TestMarketplace(t *testing.T) {
 
 	dir, err := os.Getwd()
 	require.NoError(err)
-	pluginPath := filepath.Join(filepath.Dir(filepath.Dir(dir)), "marketplace")
+	pluginPath := filepath.Join(filepath.Dir(filepath.Dir(dir)), "appregistry")
 
-	env.Must(env.Exec("add marketplace plugin globally",
+	env.Must(env.Exec("add appregistry plugin globally",
 		step.NewSteps(step.New(
 			step.Exec(envtest.IgniteApp, "app", "install", "-g", pluginPath),
 		)),
@@ -35,11 +35,11 @@ func TestMarketplace(t *testing.T) {
 	assertGlobalPlugins(t, []pluginsconfig.Plugin{{Path: pluginPath}})
 
 	listOutput := &bytes.Buffer{}
-	env.Must(env.Exec("run marketplace list",
+	env.Must(env.Exec("run appregistry list",
 		step.NewSteps(step.New(
 			step.Exec(
 				envtest.IgniteApp,
-				"marketplace",
+				"appregistry",
 				"list",
 			),
 			// all test outputs are going to the stdErr for no reason, but
@@ -49,16 +49,16 @@ func TestMarketplace(t *testing.T) {
 		)),
 	))
 	gotList := listOutput.String()
-	require.True(strings.HasPrefix(gotList, "📦"), "unexpected output: %s", gotList)
+	require.True(strings.Contains(gotList, "appregistry :"), "unexpected output: %s", gotList)
 
 	infoOutput := &bytes.Buffer{}
-	env.Must(env.Exec("run marketplace info",
+	env.Must(env.Exec("run appregistry info",
 		step.NewSteps(step.New(
 			step.Exec(
 				envtest.IgniteApp,
-				"marketplace",
+				"appregistry",
 				"info",
-				"github.com/ignite/apps",
+				"explorer",
 			),
 			// all test outputs are going to the stdErr for no reason, but
 			// it's ok when we run the app. The output goes to stdout.
@@ -67,7 +67,7 @@ func TestMarketplace(t *testing.T) {
 		)),
 	))
 	gotInfo := infoOutput.String()
-	require.Contains(gotInfo, "Description:\tIgnite Apps")
+	require.Contains(gotInfo, "Description:\tEasy to use terminal chain explorer for testing your Ignite blockchains")
 }
 
 func assertGlobalPlugins(t *testing.T, expectedPlugins []pluginsconfig.Plugin) {