+++ spaceship/cmd/ssh.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"context"
+
 	"github.com/ignite/cli/v28/ignite/services/plugin"
 )