+++ rollkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Rollkit App Changelog
 
+## [`v0.2.1`](https://github.com/ignite/apps/releases/tag/rollkit/v0.2.1)
+
+* [#106](https://github.com/ignite/apps/pull/106) Improve bonded tokens setting.
+
+## [`v0.2.0`](https://github.com/ignite/apps/releases/tag/rollkit/v0.2.0)
+
+* [#92](https://github.com/ignite/apps/pull/92) Add `rollkit init` command.
+
 ## [`v0.1.0`](https://github.com/ignite/apps/releases/tag/rollkit/v0.1.0)
 
 * First release of the Rollkit app compatible with Ignite >= v28.x.y